+++ src/tensor/mod.rs
@@ -1,6 +1,7 @@
 mod autograd;
 mod numeric;
 mod raw_tensor;
+mod rc_tensor;
 mod tensor_like;
 mod tensor_view;
 mod utils;
@@ -8,6 +9,7 @@ mod utils;
 pub use autograd::Derivative;
 pub use numeric::*;
 pub use raw_tensor::*;
+pub use rc_tensor::*;
 pub use tensor_like::*;
 pub use tensor_view::*;
 pub use utils::*;
